requirements.txt
streamlit
pandas
requests
numpy
streamlit-js-eval
supabase
google-cloud-storage